backend/requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.13.0
asgiref==3.11.1
attrs==26.1.0
billiard==4.2.4
celery==5.6.3
certifi==2026.2.25
click==8.3.2
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
distro==1.9.0
Django==6.0.4
django-environ==0.13.0
django-filter==25.2
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
docstring_parser==0.18.0
drf-spectacular==0.29.0
factory_boy==3.3.3
Faker==40.15.0
groq==1.2.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
inflection==0.5.1
iniconfig==2.3.0
jiter==0.14.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kombu==5.6.2
packaging==26.1
pluggy==1.6.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
pydantic==2.13.2
pydantic_core==2.46.2
Pygments==2.20.0
PyJWT==2.12.1
pytest==9.0.3
pytest-django==4.12.0
python-dateutil==2.9.0.post0
PyYAML==6.0.3
redis==7.4.0
referencing==0.37.0
rpds-py==0.30.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.5
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
tzlocal==5.3.1
uritemplate==4.2.0
vine==5.1.0
wcwidth==0.6.0
